project complition

+++ todolist_app/views.py
@@ -4,26 +4,33 @@ from .models import TaskList
 from .forms import TaskForm
 from django.contrib import messages
 from django.core.paginator import Paginator
-# Create your views here.
+from django.contrib.auth.decorators import login_required
+
+@login_required
 def todolist(request):
    if request.method == 'POST':
       form=TaskForm(request.POST)
       if form.is_valid():
-         form.save()
+         instance=form.save(commit=False)
+         instance.manage = request.user
+         instance.save()
       messages.success(request,'Task Added Successfully !!!') 
       return redirect('todolist')
    else:
-      all_task=TaskList.objects.all()
+      all_task=TaskList.objects.filter(manage=request.user)
       paginator = Paginator(all_task,5)
       page = request.GET.get('pg')
       all_task = paginator.get_page(page)
       return render(request,'todolist.html',{'all_task':all_task})
-
-def delete_task(requst,task_id):
+@login_required
+def delete_task(request,task_id):
    task=TaskList.objects.get(pk=task_id)
-   task.delete()
+   if task.manage == request.user:
+      task.delete()
+   else:
+      messages.error(request,'Access Restricted!!, You Are Not Allowed!!') 
    return redirect('todolist')
-
+@login_required
 def edit_task(request,task_id):
    if request.method == 'POST':
       task=TaskList.objects.get(pk=task_id)
@@ -36,17 +43,23 @@ def edit_task(request,task_id):
       task_obj=TaskList.objects.get(pk=task_id)
       return render(request,'edit.html',{'task_obj':task_obj})
    
-
-def complete_task(requst,task_id):
+@login_required
+def complete_task(request,task_id):
    task=TaskList.objects.get(pk=task_id)
-   task.done= True
-   task.save()
+   if task.manage == request.user:
+      task.done= True
+      task.save()
+   else:
+      messages.error(request,'Access Restricted!!, You Are Not Allowed!!') 
    return redirect('todolist')
-
-def pending_task(requst,task_id):
+@login_required
+def pending_task(request,task_id):
    task=TaskList.objects.get(pk=task_id)
-   task.done= False
-   task.save()
+   if task.manage == request.user:
+      task.done= False
+      task.save()
+   else:
+      messages.error(request,'Access Restricted!!, You Are Not Allowed!!') 
    return redirect('todolist')
 
 def index(request):
